Small tweak to forgeBSgenomeDataPkg()
forgeBSgenomeDataPkg() now evaluates the R expression passed thru
the 'seqnames' field of the seed file, and injects the deparsed
result (always a character vector) in the R/zzz.R file of the forged
package, instead of injecting the non-evaluated expression. For example,
if 'seqnames' is set to

  GenomeInfoDb::getChromInfoFromNCBI("GCA_002915635.3")$SequenceName

then forgeBSgenomeDataPkg() will now evaluate the expression (which can
be slow and requires internet access) and inject the result (a character
vector of length 27157) in the forged package, instead of storing the
expression itself.

With this change the seqnames are computed only once (at forge-time)
and won't need to be computed again each time the forged package gets
loaded. This ensures that loading the forged package is fast and reliable.